Fix 1.0

+++ src/com/siddhartha/clases/GrafoAciclico.java
@@ -2,18 +2,18 @@ package com.siddhartha.clases;
 
 public class GrafoAciclico {
 
-    private final int[][] mAdyacencia;
+    private final double[][] mAdyacencia;
 
     public GrafoAciclico(int pNodos) {
-        mAdyacencia = new int[pNodos][pNodos];
+        mAdyacencia = new double[pNodos][pNodos];
     }
 
-    public void AdicionarArista(int pNodoInicio, int pNodoFinal, int pPeso) {
+    public void AdicionarArista(int pNodoInicio, int pNodoFinal, double pPeso) {
         mAdyacencia[pNodoInicio][pNodoFinal] = pPeso;
         mAdyacencia[pNodoFinal][pNodoInicio] = pPeso;
     }
 
-    public int ObtenerAdyacencia(int pFila, int pColumna) {
+    public double ObtenerAdyacencia(int pFila, int pColumna) {
         return mAdyacencia[pFila][pColumna];
     }
